requirements.txt
email-validator>=2.2.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
google-genai>=1.24.0
gunicorn>=23.0.0
hyperbrowser>=0.47.0
markdown>=3.8.2
markupsafe>=3.0.2
psycopg2-binary>=2.9.10
pypdf2>=3.0.1
python-dotenv>=1.1.1
sift-stack-py>=0.7.0
sqlalchemy>=2.0.41
trafilatura>=2.0.0
werkzeug>